Fix : prevent from creating new Picks and Classifieds when @showloc is active.
Note : I had tried to add the following code to menu_picks_plus.xml :
      <menu_item_call.on_enable
       function="Picks.Plus.Enable"
       userdata="new_classified" />
to make the "New Classified" menu item be disabled under @showloc. It worked well... until you clicked on either disabled item of the menu, making the menu disappear and you wouldn't be able to make it reappear until you relogged. So I removed that piece of "code". The "New Classified" menu item is therefore always enabled, but clicking on it while @showloc is active does nothing.