Fix: could not @attach an item located inside a nested temporary folder (ex: "#RLV/~a/b/c") Fix: folder "~a/b/c" located under #RLV became "#RLV/~a/b/c" at the next login (*) Fix: declining inventory offer "#RLV/~a/b/c" created the folders in #RLV anyway instead of in the trash folder
(*) The fix is not 100% certain as we call category->updateServer() which uses AISv2 and not v3 so it seems to be a bit random, maybe subject to a race condition. For this reason we call category->updateServer() once after each rename, then a second time once the treatment is done, just to be sure.