Fix but not really a fix : When certain commands with parameters are received, for example @camdistmax, remove the instances of the same command (but with different parameters) from the list that were sent by the same object. That way sending @camdistmax:5 then @camdistmax:10 from the same object will restrict the camera within 10 m and not 5 m like before (since RLV blends both by choosing the minimal value).
It is not really a fix though because this is how the RLV is supposed to work. It retains all the commands that are sent and does not remove any until told to remove it. In this present case though, it makes more sense to overwrite existing parameters because of the blending that occurs so we're not stuck with the minimal value all the time if we want to replace it with a bigger one.