Fix some refs that are used when linking SLPlugin
/usr/bin/x86_64-linux-gnu-ld: cannot find -lmeshoptimizer.o
/usr/bin/x86_64-linux-gnu-ld: cannot find -liconv
/usr/bin/x86_64-linux-gnu-ld: cannot find -ljson_linux-gcc-4.1.3_libmt
/usr/bin/x86_64-linux-gnu-ld: cannot find -lboost_signals-mt-x64

Meshoptimizer is already a prebuilt package too on Linux, as a static
library, so we can use the same reference as Darwin's.

libiconv seems to be part of libc6 Debian, or Debian-derived, package as
a GNU standard C library, that there are no separate libiconv.* files.

The version on the prebuilt JsonCpp static library file has been
updated to 4.6.

Among the Boost libraries that must be, or can be, built separately to
be library files instead of header-only, Boost Signals is not one of
them.