Turn off viewer stat packets logging on Linux
Temporary workaround to dodge unused value warning treated as an error
by GCC, so we can have a successful build sooner. Later this commit
should be reverted, and replaced with such a warning supression, which
isn't done now because it would cause a whole section (newview, or even
worse, all if the change is applied to llcommon) needing a recompile, as
the change would need to be applied to CMakeLists.txt, or the
build-variables repo. Either way, a change to an environment value would
cause a recompile for all affected files.