This patch makes the worldmap use more accurate avatar positions in nearby regions and falls back to the old method for avatars in regions that are further away. The feature is controlled by the setting "KokuaAccurateAvatarPositions" which by default is set on. There is currently no possibility to control this setting from the UI other than through the Debug Settings and we need to decide where we want to put it. As there is no (popup-)menu in the worldmap it will probably need to be added somewhere in the preferences.
This patch also improves performance in the LLWorld::getAvatars() function by moving certain conditions higher up into outer loops or outside of the loops and aborting loops as soon as possible when doing filtering.

This patch also does replace NULL with nullptr in the relevant files and attempts to remove unused headers.